UI changes
UI_Manager
- Added RestartScene() in UI_Manager, makes GameController RestartScene obsolete
- Made LoadScene(), load the next scene automatically and give a warning if no scene was inputted in the field
- Made a BackToMainMenu(), which returns to Main Menu Scene

UI
- Added Return to main menu button, which returns to mainmenu using UI_Manager
- UI Prefab should now just work when dropping it in a scene. ONLY CHANGE UI IN THE PREFAB!

+++ mato/Assets/Scripts/Game_Controller.cs
@@ -16,7 +16,7 @@ public class Game_Controller : MonoBehaviour, IUnityAdsListener
     public Canvas loseUI;
 
     // Amount of points needed to win
-    public float winCondition = 3;
+    public int winCondition = 3;
     public bool play = true;
 
     private float secondsElapsed = 0;
